+++ line/line_segment.h
@@ -4,18 +4,18 @@
 
 class Line_segment_t final {
 private:
-    float t1;
-    float t2;
+    double t1;
+    double t2;
 
 public:
     Line_segment_t(Triangle_t& triangle, Plane_t& plane, Line_t& line);
     
-    float get_t1() const;
-    float get_t2() const;
+    double get_t1() const;
+    double get_t2() const;
 
-    float compare_segments(Line_segment_t line_segment) const;
+    double compare_segments(Line_segment_t line_segment) const;
 
 private:
-    float intersection_point(Point_t vertex1, Point_t vertex2, 
-                             float dist1, float dist2, Line_t& line) const;
+    double intersection_point(Point_t vertex1, Point_t vertex2, 
+                             double dist1, double dist2, Line_t& line) const;
 };
